Souncloud - 1st commimt
Look up, and open songs from soundcloud.

I do not include the artist in as the summary on purpose, it makes the
workflow twice as slow and I find it untolerable. If you would prefer
to deal with the lag and see the artist under summary, use the code
from SearchSoundCloud-ALT.py as your script filter, and set link it to
the action "open url" instead of "run script"